Update 11/8/2016 end of meeting

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/DriveTrainInterface.java
@@ -9,7 +9,7 @@ public interface DriveTrainInterface
 {
     void init(HardwareMap hardwareMap);
     void drive();
-    void driveStraight(double Distance, int degrees);
+    void driveStraight(double Distance, double degrees);
     void spin(float degree);
 }
